Fix list of cached docker images not getting updated
by using the ephemeral cache as it was intended.

In https://github.com/galaxyproject/galaxy/pull/10451/ we started
passing in the beaker-backed persistent cache, when the code was
expecting to be able to set per-request ephemeral values in the
ResolutionCache object. This fixes new docker images not being returned
in the CachedDocker resolver.

That is the simplest fix, but we should probably refactor this in dev so
that even successfull resolution has an expiry date, in case there is an
updated image on quay that should be used.