implement a 'duplicate_column_filter'
assume there is a data table with 2 columns: name and value
(e.g. mothur data tables and probably many other old ones).

tools using such data tables can not be tested (to the best of my
knowledge) since one needs to specify the value ..
which is a path that changes with the test environment.

my idea is to add a filter that duplicates a column, i.e.

(name, value) can be transformed into (name, name, value)

this already works, but only for accessing the columns by
index. accessing the columns by name does not work, since
the names still refer to the old columns (which might be
confusing for the tool developer .. at least it deserves
proper documentation)

alternatively one might implement this in `DynamicOptions`
by adding a new attribute.

Also adds the possibility to access SelectToolParameterFieldWrapper
by index. Which was already documented, but did not work since
attributes need to be strings.