Replace StreamBall, replace tarballs with zip
The way StreamBall writes to the wsgi response just doesn't work for
ASGI frameworks. The zipstream method seems superior in that it creates
a uncompressed zip archive and doesn't need a tempfile.
Additionally I think it makes sense to return zip files, as that is
probably the most well known compression format, and we could replace
this conditionally with
https://www.nginx.com/resources/wiki/modules/zip/ (while we also work
on creating archives in a celery worker).
zipstream was one of the suggested libraries in https://github.com/galaxyproject/galaxy/issues/9794#issuecomment-631425320