Simplify JobInformation and DatasetInformation components
JobInformation contained a lot of dataset related infos.
This reworks the JobInformation component to
- only deal with job-related fields
- pushes the off-topic fields into DatasetInformation
This also adds a component that can nicely display rows with
code (stdout/stderr/command line) and that can expand these.
I also refactored out the id decoding logic into a standalone component
for re-use.

This will be very useful for the invocation PR.