Fixes for license and creator saving and updating.
I was mutating a property, it was possible to save the workflow and lose these, etc..

I was emulating annotation handling - which is an anti-pattern and should probably be redone to do what I'm doing here. This approach is much cleaner and gives Index.vue the single source of truth (and simply reactive truth) on these properties - which is nice and needed in some downstream work.