Drop unreachable showWarnings method in workflow editor.
It was added in https://github.com/galaxyproject/galaxy/commit/c922711612f217d0d780581b3ecaf5cfa3397f77 which didn't have the exception handler. showWarnings only shows a modal if the response is JSON with 'errors' in it, I've searched the model/__init__.py, managers/workflows.py, api/workflows.py, and controller/workflow.py and the only response that would include that is in the save_workflow_as. the saveAs function though does not use this function.