Drop 'unreachable'-ish showWarnings method in workflow editor.
It was added in https://github.com/galaxyproject/galaxy/commit/c922711612f217d0d780581b3ecaf5cfa3397f77 which didn't have the exception handler for this request that is now there - so I think it much have been responding to different API requests.

Regardless of history, ``showWarnings`` only shows a modal if the response is a JSON object with the key ``errors`` in its root. I've searched the model/__init__.py, managers/workflows.py, api/workflows.py, and controller/workflow.py and the only response that would include that is in the ``save_workflow_as``, but the ``saveAs`` function in the client does not use this ``showWarnings``.