Fix API tests failing when usegalaxy.org is down
`site_down` was broken when a `requests.get()` raises a timeout exception
(should have returned `True`), so I've inverted the logic to check if
the site is up, which is more natural.

Added a proper test for uploading an invalid URL, renamed the one which was
actually testing a 404 HTTP error.