Add missing id keys to built-in tours
Pydantic models (currently added to tours) require explicit
specification of model fields. Having an "id" field as required seems
right from a design standpoint.

However, if there is a reason for an "id" value to be included in the
galaxy_ui tour but not in the other tours, this field can be made
optional: `id: Optional[str] = None`