Don't attempt to use docker cli if it is not available
This fixes https://github.com/galaxyproject/galaxy/issues/11125, which
became a problem with https://github.com/galaxyproject/galaxy/pull/10814
and the recent removal of the docker runtime from kubernetes.

This is the quick fix, but I don't like the guessing and disabling
that we do if docker is not available or not functional.

I'm hesitating between adding a config option for the resolver
(`docker_cli_available` or `docker_list_images` + `docker_pull_images`)
and adding a `KubernetesContainerResolver` / `RemoteContainerResolver`
class that doesn't assume docker is available for listing and pulling
iages.

Not wanting to list/pull images on the Galaxy node is probably a
reasonable thing outside of kubernetes as well, so I tend to
go with a `RemoteContainerResolver` class in dev ?