Fix populated state for empty collections
This matches the non-optimized variant. If there's an empty collection
(because nothing has been discovered for instance)
DatasetCollection.populated would be true and
DatasetCollection.populated_optimized would be false. This is because of
the way we build the query. Noticed this with a workflow that worked
before https://github.com/galaxyproject/galaxy/pull/10917, which
switched the inputs ready check for database colletion tools to use the
populated_optimized method.