lint for optional or muliple selects with radio display
select with `display="radio"` one can not be unselected once an options was chosen.
also radio style selects do not allow to select multiple options.

both might as well be considered bugs. but I think the idea of radio buttons could
as well be considered incompatible with multiple=true or optional
https://en.wikipedia.org/wiki/Radio_button