Replace starlette-context with BaseHTTPMiddleware
this is still not ideal, as the entire response will be loaded into
memory. This is a problem with streaming responses.
https://github.com/encode/starlette/issues/1012#issuecomment-673461832
is actually enlightening here:

> "This means this class will either load the entirety of streaming requests into memory (this issue) and run the background before returning the response (#919 ), or if we fix those problems, that it will then encourage users to leave resources in a pending or open state, an arguably worse result. In short, it's problematic."

Which is exactly what we'd be doing with a middleware, keeping resources
open for longer than necessary, which we need to avoid if we ever want
to run background / async tasks.  I think the solution here what we
already had, path operation dependencies.