Fix conditional and section parameters becoming runtime parameters
This broke in 526fb05d5c7d0e154f7df2f622045059105451e5, which fixed
cloned data inputs missing from the tool form.
The problem with this commit is that we now add the runtime value state
to non-data hidden inputs, which is not correct and breaks execution
of workflows with these inputs.

This fix is better since it does not change the input type, and instead
adds a new attribute. The tool form and its variants need to be
rewritten in vue ASAP, so we can stop these hacks.