Fix datatye-defined column name display
I think this was the intention when checking for None, but this doesn't
work when the default type is not None, such as a list for column_names.
Fixes https://github.com/galaxyproject/galaxy/issues/11443

This might be another side-effect of
https://github.com/galaxyproject/galaxy/commit/335a89772012041538e86835a4e16f3a2e50d51f,
which is of course the more correct thing to do.