Break up dynamic dataset collection into chunks
Addresses https://github.com/galaxyproject/galaxy/issues/11488.

One drawback here is that if multiple jobs are concurrently creating
datasets in a history the HID is not guaranteed to be sequential
between chunks. We can address this, but this is also a problem
when breaking workflow executions at
``maximum_workflow_jobs_per_scheduling_iteration``.
If we consider this to be a problem (personally I don't think it is, but
I could be convinved otherwise) I would work on this in a different PR.

On the upside the history panel will display an increasing amount of
datasets created during the dataset discovery.