Do not use the subqueryload option
This appears to be creating a redundant join.
The tool dependency is already eagerly loaded and the subqueryload joins
the same tool_shed_repository that is already in the query.

This causes a bug under SQLAlchemy 1.4: the deferred option is not
applied to repository.metadata.

(This reverts https://github.com/galaxyproject/galaxy/pull/8933)