Restore setting hdca.update_time
and fix the subcontainer query to filter the update time on the hdca,
not on the DatasetCollection. This is necessary for HDCA state changes
in the new history panel.
Using the `with_for_update` should work around the observed deadlocks,
and `skip_locked` should prevent any slowdown when many processes
finish jobs belonging to the same implicit collection.