Use explicit base class for declarative mapping (see note)
This solves the issue of the legacy `table` attribute of mapped classes.
With declarative mapping, the generated Table is accessible via the new
`__table__` class attribute. While it is possible to change most cases
of [class].table to [class].__table__, there are a few cases when the
[class] is a variable (grids.py is one such case); so until all classes
are migrated to declarative, both `table` and `__table__` will be
present.

It is possible to solve this within grids.py (via adding a
get_model_table function or method). However, a clearner way is to use
the __delcare_last__ hook on an explicitly constructed Base class, where
we define the additional table accessor `table`. An explicit Base class
may be also useful for further refactoring of our very verbose model.

The Base class is used for all but one classes: using it for Quota
raises an error, triggered by a mapping statement. I am not yet sure why
this happens, but I'll clean this up before this PR is ready.