DefaultQuotaAssoc should not be a subclass of Quota
As far as I can tell, there is nothing in the codebase that suggests
that DefaultQuotaAssociation class should be a subclass of Quota.

This raised an error when Quota was mapped declaratively using
a Base class, with DefaultQuotaAssociation mapped imperatively.
(The error was not raised when Quota was mapped via declarative
decorator.)