Set cache_ok=True for custom types
Removes SAWarning.
Cause of warning: cache_ok added in SA 1.4.14. Default is None, which
generates a warning and does NOT allow caching of a statement that
includes this type. If it is safe to be used as part of a cache key, set
cache_ok = True.

Ref:
https://docs.sqlalchemy.org/en/14/core/custom_types.html?highlight=cache_ok#sqlalchemy.types.TypeDecorator.cache_ok
(also explains criteria for safe)
Example of not safe for cache: https://github.com/sqlalchemy/sqlalchemy/issues/6436