Add switch for optional select in workflow editor
If the `Set value for this optional select field ? `switch is set to no
we serialize the value as `null` instead of `""`, which is especially
important in optional select fields, where `""` may correspond to an
option that is distinct from not setting a value.

I have only enabled this for select parameters that are cast to text
parameters, but I think this discrimination between set and unset
would also be great for optional text parameters.
There might be a better way to do it in the UI (maybe some icon that
indicates that the value is set, that can be toggled with an backspace
when the field is empty?), but I think this will do for the relatively
rare case of optional select parameters that are cast to text fields.

Fixes https://github.com/galaxyproject/galaxy/issues/12081