Fix persisting tool state when invoking workflow
If the `tool_inputs` column of the `workflow_step` is loaded as
MutableJSONType any "accidental" change of values in the dict marks the
object as "dirty" in the sqlalchemy session it is attached to and will
be persisted when the session is being flushed.
Making the column type `JSONType` instead of `MutableJSONType` means
that the object is not marked as dirty and will not be persisted in that
case. The correct method to use to actually persist step inputs is
WorkflowModule.save_to_step or setting step.tool_inputs directly.  This
fixes https://github.com/galaxyproject/galaxy/issues/9886 and
https://github.com/galaxyproject/galaxy/issues/12081#issuecomment-860130404

The issue in https://github.com/galaxyproject/galaxy/issues/9886 is that
when you submit a workflow the step state is converted to a tool state
in
https://github.com/mvdbeek/galaxy/blob/90dff9af5bec8bed1258620e5351d0350e284206/lib/galaxy/workflow/modules.py#L661:
```
    def step_state_to_tool_state(self, state):
        state = safe_loads(state)
        if "format" in state:
            formats = state["format"]
            if formats:
                formats = ",".join(listify(formats))
                state["format"] = formats
        state = json.dumps(state)
        return state
```
and safe_loads just returns the same object if the object that is passed
in is not actually a JSON string. So `state["format"] = formats`
manipulates the step state in place and that gets committed when
flushing the new invocation.