Wait for uploaded datasets in API tests
Most methods that stage datasets already do that anwyway. Might be nice
in the future if we could just cancel running jobs at the end of a test
and not wait for uploads where not necessary, but this should increase
test stability in the short term.

Fixes https://github.com/galaxyproject/galaxy/pull/12136#issuecomment-861531075