Fix failure to serialize invocation
Old imported subworkflows may not have a StoredWorkflow instance, so
this was causing AttributeError if you're listing invocations for such
an invocation. But we also agreed to remove this in
https://github.com/galaxyproject/galaxy/pull/11881#discussion_r618314707
so I'm doing that now. With the release being just a few days old I hope
we can still make that change.