Fix bug (see note) + map HDARatingAssoc declaratively
BUG:
In the previous version, HDARatingAssoc was never set.

The attribute that was set was `history_dataset_association`. However,
this value was never translated by SQLAlchemy into a relationship
because there is not connection between this attribute and the
`history_dataset_association_id` foreign key. Thus, a relationship was
NOT added upon session flush.

The relationship that was defined was `hda` - defined in mapping.py
under HistoryDatasetAssociation mapping via backref attribute.

But this is covered by a unit test in
test/unit/data/test_galaxy_mapping.py: why did the test pass?
Because the test was NOT doing what the code does; instead the test was
using `setattr` to set the correct attribute in the test itself. So the
test avoided the bug via an alternative implementation of the code under
test.

In my option, the following will help avoid such issues:
1. Use back_populates instead of backref to keep relationships explicit.
2. Keep testing code simple.