quote extract_dataset_parts.sh call
In https://github.com/galaxyproject/galaxy/pull/12104 I realized that
`<pararalellism>` does not work if the galaxy root dir contains a space.

This would fix it, but maybe (probably) there is a better way do do it
(galaxy root should always be quoted).

No idea how this was not realized earlier since the converter tests
use a space in galaxy root for a long time...