Make conda_auto_init option overrridable
When using galaxy-k8s as a container for testing a single instance of galaxy, it is very useful if that galaxy can install tools. Having `conda_auto_init` on `False` is desirable when running kubernetes, but for other use cases it might be better to have it on `True`. Currently that requires mounting a new galaxy config file in the container (also overwriting the default configuration settings), but with this change the setting can be overridden with an environment variable. This is much more convenient.