show target datatype in datatype conversion
in https://github.com/galaxyproject/galaxy/pull/12185 we introduced
that a single converter can convert to multiple target datatypes.

the current conversion form shows only the converter names in the
select form. this is confusing since all targets have the same name.

with the change the form shows the target datatype and in parenthesis
the name of the converter.