Map LibraryDataset declaratively (see note)
If we have:
`Column('info', Sting, key='_info')`

then we remap like this:
`_info = Column('info', String)`

Because the key attribute will be overridden. In both cases, we have
access to LibraryDataset._info (assuming it's an instance of LibraryDataset).

In the case of LibraryDataset, if we were to do this:
`info = Column('info', String, key='_info')`, which is the intuitive way
to do it, then we'd get an error when we try to access
LibraryDataset.info BEFORE explicitly setting it. `_info` is NOT created
by the mapper, so when we access the property `info`, the code first
checks the value of `self._info` - which will raise an Attribute Error.