deprecate data manager version
Data managers (DM) could define a version for the data manager
(`<data_manager version="0.0.1">`) and for the data manager tool (DMT).

The version attribute of the data manager was undocumented and only
used seldomly (for IUC in 24 out of 88). If no version was set for
the DM it defaults to 0.0.1. Hence in `shed_data_manager_conf.xml`
multiple versions of the (DM) with version 0.0.1 are listed.

This leads to loading only one DM (I guess the oldest) and a
warning like the following on startup:

`A data manager has been defined twice: toolshed.g2.bx.psu.edu/repos/galaxyp/data_manager_eggnog_mapper/data_manager/data_manager_eggnog/0.0.1 `

which is easily overlooked.