Add fixture factories
In some tests we may need more than one instance of the same model. We cannot reuse a model
fixture, and we cannot pass multiple copies of the same fixture to one test. We have to
instantiate a new instance of the model inside the test. However, a test should only know
how to construct the model it is testing, so instead of constructing an object directly,
a test calls a factory function, passed to it as a fixture.