Do not set SQLAlchemy state on a wrapped object; undo 0e3afe7
Here's what happens when an instance of SafeStringWrapper is created.
(self = instance of SafeStringWrapper)
A class mapped declaratively is a subclass of DeclarativeMeta. It will check at creation time
if self has _sa_instance_state set, and if not, it'll try to set it. This happens BEFORE self.__init__
has been called, so self.unsanitized does not exists, which raises an AttributeError.
To avoid this, as well as to avoid SQLAlchemy state to be set on SafeStringWrapper,
we simply ignore this call.