Fix constructor inheritance bug (see detailed note)
MetadataFile subclasses StorableObject, which defines a constructor.
That constructor was called from MetadataFile's constructor via
`super().__init()`.

MetadataFile also subclasses Base, which now defines its own constructor
(after the fix in the previous commit). As a result, the call to super()
resolves to the constructor defined on Base, NOT on StorableObject. This
results in self.uuid not being defined - and that kills everything.

One approach would be to exploit how method resolution order operates
and simply switch Base and StorableObject in the class definition - but
that's a smelly hack; besides that would mean that SQLAlchemy's default
constructor would have been made inaccessible.

A better solution is not to have a separate constructor just for the
sake of setting the uuid (especially considering the same logic is
duplicated in 3 more places!) and, instead, factor out that logic into a
helper function, and call that function from the constructors of both
objects that subclass StorableObject (there are only 2). Which is what
this code does.