Remove deprecated workflow run path through the API.
It has been marked as deprecated in Galaxy since 2015 and from bioblend since 0.7.0 (also released in 2015).

After this is removed /api/workflows will only created StoredWorkflows - this will make it a lot easier to write documentation during the FastAPI transition. This also allows us to remove a bunch of hacks around conditionally invoking workflows different ways - both in testing and in the implementation.