Drop Job.job relationship + fix bug (see note)
Initially, the Job.job relationship was created implicitly via backref,
which was incorrect. However, this attribute masked a bug in
workflow/run.py - which is fixed in this commit.

BUG DESCRIPTION: the job_assoc is a Job, not a Job[Foo]Association -
which should have a job attribute. A Job should not have a job
attribute.

The bug was exposed by running the
test_workflows.py::WorkflowsApiTestCase::test_run_with_implicit_connection
test, which populates step_invocation.jobs with a nonempty list of jobs
(not job associations).