Delay start of job runners
Fixes https://github.com/galaxyproject/galaxy/issues/12570.

The problem was that the JobDispatcher also starts the job runners'
`_init_worker_threads` method, which starts the queue worker threads.
This is guarded by `register_postfork_function`, but the standard
implementation (for anything that is not uwsgi pre-fork) just executes
the function right away.

We could alternatively implement a standard register_postfork_function
that we could use to delay functions until after the app is built
correctly. Not sure what is preferable, but this change should be fairly
minimal.