Fix wrapped_class_name bug (see note)
At this point in the code, we may have a CallableSafeStringWrapper. This
is unlikely, but we provide for this possiblity by referring to the
class using the safe_class variable when creating the new type a few
lines below.

An unlikely use case of a CallableSafeStringWrapper is covered by a unit
test (instance of a class that has the __call__() method defined)