only filter datasets when dbkey is set
Like in the filter for selects with dynamic options.
Note actually in the select case we check if the referred
dataset has the dbkey set while we do this here for the
dataset itself. I guess this could be changed but I could
not find a good way to do this.

Discovered while fixing https://github.com/galaxyproject/tools-iuc/pull/4023
for https://github.com/galaxyproject/galaxy/pull/12073

where this leads to a bug because filtered datasets are not matching:
https://github.com/galaxyproject/galaxy/blob/5ea4d7e647d09dde957e824a083b1bd6acaa6f17/lib/galaxy/tools/parameters/dataset_matcher.py#L132