Fix RepositoryMetadata review/reviews. For real this time (see note)
There were 2 relationships defined on RepositoryMetadata: review and
reviews. The former (review) must have been added unintentionally via
backref. It was not referenced in the codebase, and its name was
incorrect too: that relationship holds a list of reviews.

However, the foreign_key attribute was CORRECT, whereas the correct
relationship (reviews) had an incorrect foreign_key attribute, which
resulted in an incorrect set of reviews (the size of the set was at most
1, which is incorrect). For an example of correct usage, see
https://docs.sqlalchemy.org/en/14/orm/relationship_api.html?highlight=foreign_keys#sqlalchemy.orm.relationship.params.foreign_keys

This commit corrects the foreign_keys attribute and fixes the unit test
that verifies correct behavior.

Victory at last!