Move RepositoryMetadata out of mapping.py
Do not remap declaratively. The model contains '.metadata' attr; a
declaratively-mapped class cannot have a .metadata attribute (it is used
by SQLAlchemy's DeclarativeBase). Given that TS is currently (mostly) in
maintenance code, it is reasonable to leave this as is rather than
handle all references to this attribute in the codebase.