Framework support for integration tests requiring celery.
Rework task decorators to be less verbose and so we don't need hacky delay-ed import of the task functions and so there are no circular dependencies between galaxy.celery.__init__ and galaxy.celery.tasks. Not to say the get_galaxy_app() in there isn't its own kind of hacky, but I think on balance this is a lot cleaner.