Fix virtual scroller missing item updates
I don't fully understand what is going on, but here's my current
diagnosis of why this is ncessary:
Rhe contents/near response contains the max_hid header, and the scroller
uses that to determine if we're updating at the top of the scroller,
[here](https://github.com/mvdbeek/galaxy/blob/90c797bc485e238d9bfd0bd55b1b7c5f978037a7/client/src/components/History/providers/HistoryContentProvider/contentPayload.js#L153.)
For some reason `updatesAtTop` is false when the test fails because both
`response.maxHid` and `lastResponse.maxHid` are at 2. If I change > to >=
the test consistently passes in 10 tries, while the original fails in 5
out of 10 tries. I guess pairwise only starts emitting once there are
actually 2 responses so if the initial response contains both items (hid
1 and hid2) I guess it'll fail.