Fix wrong dataset details link
Fixes https://github.com/galaxyproject/galaxy/issues/12804.
This is not the ideal fix, the client should ideally contruct the link
... but it's the old history, it's an easy fix and it's consistent with
what the new history does (for now).

There is some backbone manipulation of the model id (which is not
HistoryDatasetAssociation, but a DatasetCollectionElement in
a collection context)  which is supposed to set the id to the
HistoryDatasetAssociation id. I guess it may not have run on time ?
I also don't know where it is and how to find it, might have been
deleted ? I think it was related to
https://github.com/galaxyproject/galaxy/blob/dev/client/src/mvc/collection/collection-model.js#L86.