Fix response model issues
Enabling this flag will avoid returning fields with default values.

The `type_id` field was not returned by the services dictionary response but the pydantic model was adding it with the default value of None. This was causing the client to behave in a strange way when editing the name of a collection, making a POST request to create a new collection instead of making a PUT request with the name change.