Fix OpenAPI docs not rendering
There is a regression bug in FastAPI/Pydantic (see https://github.com/tiangolo/fastapi/issues/3665) that prevents the OpenAPI interactive documentation to render when we have models using `Tuple`.
Since Tuple is currently used just in one place, we can temporarily replace it with a simple List, we will lose a bit of validation but still is not worth reverting back to a previous version of FastAPI because of this.