Avoid setting response headers through `trans` in datatypes
The `display_data` function access `trans.response` to set or get the response headers for different kinds of response contents.
The `trans.response` is no longer present in FastAPI mode (unless we decide to merge #12929 which has its own issues) and, in addition, we are trying to avoid or reduce the use of god objects as much as we can.