Disable remote evaluation for JEHA history export jobs
We'd have the add `JobExportHistoryArchive` to the supported job outputs
in the model export store or move the `JobExportHistoryArchive` creation
from `execute()` ToolAction to the handler and skip creating
`JobExportHistoryArchive` altogether, but we want to perform the
job export in the Celery workers anyway, so I think it's not worth the
effort.
Note that exporting to file sources doesn't use
`JobExportHistoryArchive`, so for those we can still create the command
line remotely (and e.g materialize deferred datasets in a history).