Reuse engine in DatabaseStateCache
This adds an extra line of code to tests that directly access
DatabaseStateCache. We also store one or two engine instances in
DatabaseVerifier (which we reuse in DatabaseStateCache and, as a next
step, in AlembicManager).

Creating an engine is expensive and wasteful, so this, I think,
is a reasonable tradeoff between complexity and efficiency.