Retrieve engine from scoped_session via get_bind
For ref: https://docs.sqlalchemy.org/en/14/orm/contextual.html?highlight=scoped_session#sqlalchemy.orm.scoping.scoped_session.get_bind

I wasn't able to find a description of the bind attribute in the docs,
although it is present in the source code. It is my understanding that
it will work in most cases. However, accessing it via get_bind() ensures
that the correct relosultion order is applied (see method
documentation). So, this ensures it will work in all cases.