Ignore SA error when dropping migrate_version tbl (see note)
This is expected behavior. Normally we won't try to add or delete the
same object to/from the same database. The migrate_version table is an
exception: it exists in both models: galaxy and install model. When
dropping it, we drop it from both schemas - so we have 2 revisions
defined, and if the database is combined (as it is in at least one
test), the revisions with the same operations will be applied to the
same database.

We catch and ignore SQLAlchemy's InvalidRequstError: this handles both
sqlite and postgres use cases (OperationalError will only handle
sqlite).