Do not load db state before creating a db
Also, do not call private method of DatabaseVerifier, even in tests: now
it will break because it's supposed to work only AFTER the db state has
been loaded - which is loaded lazily and only after verify() has been called.

Instead, add a helper method to the test module + a test for the helper.