Add `homogeneous_datatype` field to API response
If all the datasets in a collection have the same extension (or datatype), the new `homogeneous_datatype` field will contain that extension indicating that the collection is 'homogeneous'. If the field is None, then the collection has mixed data types in its datasets (heterogeneous).

Currently, the field is only serialized using the 'betawebclient' view.