Don't store stdout/stderr in info field
This is pretty nonsensical given that we do this for each output dataset,
when that is a property of the job that produced this file. It just
takes up unncessary storage in the database (pretty minor to be fair, we
limit the column to 255 chars). It also happens we have a
"bug" in the extended metadata mode where we don't limit the size of the
info property.
If we want to display something like this we should build it dynamically
from the job stdout/stderr, but I don't see why we would. This information is
available in the dataset details view.