Drop tool_migration_manager and all references to it (see note)
The from_tool_migration_manager parameter as False as its default value.
The only case where it's called with the value True is when it's called
from the from_tool_migration_manager module - which is now deleted.
Therefore, now the value is always False and, therefore, it is safe to
drop this parameter and adjust the code that checks its value accordingly.