Add default files as additional input param types
This isn't really that useful yet because the editor verifies the
datatype, and we're not setting the datatype here, so you can't create
connections in the UI.
This also looks a little different from CWL workflows:
```
steps:
  step1:
    run: wc-tool.cwl
    in:
      file1:
        source: file1
        default:
          class: File
          location: whale.txt
    out: [output]
```
In CWL the default file is inline with the workflow step, while in this
implementation the default file is its own step (since all explicit
inputs are steps at this point). If we're aiming for CWL export this
could be rewritten at export time. However I don't think this helps us
with default files in CommandLineTools.

An alternative implementation that would be closer to CWL is to allow
`DataToolParameter`'s in a workflow context to behave like
`DefaultDatasetToolParameter`s, so you could select a dataset in the
workflow tool form.
The downsides:
- we would need to parse all workflow steps to find all default file
  inputs that should be overridable in the workflow run form (but to
some degree we do that already anyway ?)
- in the current UI it would be difficult to find all the default files
  of a workflow if they're "hidden" in a step

A lot of things are missing here:
- exported tool state stores decoded dataset ids
- should probably extend regular dataset parameter instead of input
  parameter
- default files should come as non-optional as well
- default files should not live in histories IMO, but need to figure out
  quota management and other things
- sharing
- user interface + API for exporting such workflows (needs to include)
- user interface + API for importing workflows with default files