fix MetadataValidator
If no check or skip is given, self.check and self.skip
need to be initialised to None. Otherwise validation with
`value.missing_meta(check=self.check, skip=self.skip)`
is called with `value.missing_meta(check=[""], skip=[""])`.

This led to the stalling observed in
https://github.com/galaxyproject/galaxy/issues/13136

Explanation: each data input parameter automatically gets
a MetadataValidator. Since this validator was dysfunctional
the job was scheduled anyway, but did not execute.