Fix test_index_using_keys
For some reason this test keeps returning {} (empty object) instead of the `size` when running on CI (but passes on local).
Maybe this `size` key gets populated after some time?
I changed the expected key to `id` and check there is only one key returned which is what the test is really about.