lower the min username length limit to 1
Two would be sufficient to make at least one `Yu`, `Li`, ... per Galaxy
instance happy :) Korea seem to allow for single letter names.

Anyway, I just want to make Mr T. happy :)

The serious background is that this seems to be an easy way to solve the
following problem: If a user is registered some transformation
of the username happens, e.g. it is filled by - to enforce the minlength
requirement. On LDAP managed systems that submit jobs as real users
this leads to the problem that users with to short names can't submit
jobs.