add function to return lint context
this adds functionality to the linter functions to return the lint
context. without breaking previous behavior.

Planemo seems to use

- lint_tool_source (which returned the fail state) and
- lint_tool_source_with (which returned nothing)

For the later a parameter has been added to make the linter silent
and the lint context is now returned.

For the former a variant has been added which returns the lint context.