Store peptide IDs as strings.
These peptide IDs were previously parsed as floats. The result was that
peptides that were too large were rounded, and the ID was no longer
meaningful, resulting in the "PSM Detail for Selected Peptides" table
not being populated.

With this change PEPTIDE_ID field is no longer parsed as a float, and
remains a string.