Backport SingleQueryProvider improvements
Use ``timeout``, not interval, bump to 3s, make stoppable

Depending on the response time using setInterval can lead to a pileup of
requests, see
https://developer.mozilla.org/en-US/docs/Web/API/setInterval#ensure_that_execution_duration_is_shorter_than_interval_frequency.

I've also bumped up the 500 ms interval to 3000ms. 500 ms is too low and
will cause high server load.

This also adds an optional `stopRefresh` function that allows the caller
to determine when to stop polling, based on the returned response,
and emits a `update:result` event that callers can use to update local
data.